Made the app a little more modular
Tried to repeat myself less. Originally I was doing the same thing
over and over with each route. So I made a class to display the main
app pages called FrontPage. (I think the individual place pages will be too
different?) I'm not sure if this is the right way to do it, but the code
looks cleaner now. That's something.